requirements.txt
# ============================================
# ORBBEC Gemini 2 工業檢測系統 - 依賴項目
# ============================================

# ----- 深度學習框架 -----
torch>=2.1.0
torchvision>=0.16.0
ultralytics>=8.0.0  # YOLOv11 支援

# ----- 3D 處理與點雲 -----
open3d>=0.17.0
trimesh>=3.21.0

# ----- 影像處理 -----
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0
Pillow>=10.0.0

# ----- 數值計算 -----
numpy>=1.24.0
scipy>=1.11.0

# ----- 資料處理 -----
pandas>=2.0.0

# ----- 配置檔案處理 -----
pyyaml>=6.0
python-dotenv>=1.0.0

# ----- 可視化 -----
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# ----- 進度條與日誌 -----
tqdm>=4.65.0
loguru>=0.7.0

# ----- API 與資料驗證 -----
pydantic>=2.0.0
fastapi>=0.100.0  # 未來可能用於 API 服務
uvicorn>=0.23.0

# ----- 測試工具 -----
pytest>=7.4.0
pytest-cov>=4.1.0

# ----- 程式碼品質 -----
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# ----- Jupyter 支援（開發用）-----
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# ----- ORBBEC SDK -----
# 注意：OrbbecSDK-Python 需要手動安裝
# 請參考: https://github.com/orbbec/pyorbbecsdk
# pip install pyorbbecsdk

# ----- TensorRT（Jetson 部署用）-----
# 注意：TensorRT 通常隨 JetPack 一起安裝
# 在 Jetson 平台上使用以下方式驗證:
# python3 -c "import tensorrt; print(tensorrt.__version__)"

# ----- ONNX 支援（模型轉換）-----
onnx>=1.14.0
onnxruntime>=1.15.0  # CPU 推理
# onnxruntime-gpu>=1.15.0  # GPU 推理（與 onnxruntime 擇一安裝）

# ----- 其他工具 -----
scikit-learn>=1.3.0
scikit-image>=0.21.0
